create_image.sh is now build.sh
This script does not create a disk image any longer

It can now be run as any user and does not require root priviledges any
more.

The create_image_helper is now called create_disk_image and in the main
directory. It requires root priviledges and needs the output directory
as argument. It will use the output directory below the current working
directory otherwise.